refactor: 테스트에 도메인 패키지 생성 및 클래스 이동
