+++ src/test/java/domain/NameTest.java
@@ -1,3 +1,5 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
